update
- added music transition time
- fixed game thread update bug
- fixed melee attack
- added ingame on exit transition
todo: added on enter and on exit transitions to game states

+++ gameProject/Game/gameObjects/gameState/inGameState/inGameState.h
@@ -18,6 +18,6 @@ private:
     void setCameraPosition( );
     bool isPaused = false;
     float zoomLevel = 0.0f;
-    int currentExitFadeAlpha = 0.0f;
+    float currentExitFadeAlpha = 0.0f;
     void exitGame( gameStateManager * manager );
 };
